+++ src/main/java/local/ts3snet/unicbot_ms_spring/module_webutils/service_ekey/entity/KeyDataEntity.java
@@ -1,4 +1,4 @@
-package local.ts3snet.unicbot_ms_spring.module_webutils.entity;
+package local.ts3snet.unicbot_ms_spring.module_webutils.service_ekey.entity;
 
 import lombok.Getter;
 import lombok.Setter;
