+++ src/main/java/local/ts3snet/unicbot_ms_spring/module_webutils/service_ekey/model/WebChatNode.java
@@ -1,4 +1,4 @@
-package local.ts3snet.unicbot_ms_spring.module_webutils.model;
+package local.ts3snet.unicbot_ms_spring.module_webutils.service_ekey.model;
 
 import lombok.Data;
