+++ src/main/java/local/ts3snet/unicbot_ms_spring/module_webutils/service_ekey/repository/KeyDataRepository.java
@@ -1,6 +1,6 @@
-package local.ts3snet.unicbot_ms_spring.module_webutils.repository;
+package local.ts3snet.unicbot_ms_spring.module_webutils.service_ekey.repository;
 
-import local.ts3snet.unicbot_ms_spring.module_webutils.entity.KeyDataEntity;
+import local.ts3snet.unicbot_ms_spring.module_webutils.service_ekey.entity.KeyDataEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDateTime;
