+++ src/main/java/local/ts3snet/unicbot_ms_spring/module_webutils/service_ekey/service/KeyDataService.java
@@ -1,6 +1,6 @@
-package local.ts3snet.unicbot_ms_spring.module_webutils.service;
+package local.ts3snet.unicbot_ms_spring.module_webutils.service_ekey.service;
 
-import local.ts3snet.unicbot_ms_spring.module_webutils.entity.KeyDataEntity;
+import local.ts3snet.unicbot_ms_spring.module_webutils.service_ekey.entity.KeyDataEntity;
 
 import java.time.LocalDateTime;
 import java.util.List;
