+++ src/main/java/local/ts3snet/unicbot_ms_spring/module_webutils/service_ekey/service/UtilsWebChatService.java
@@ -1,4 +1,4 @@
-package local.ts3snet.unicbot_ms_spring.module_webutils.service;
+package local.ts3snet.unicbot_ms_spring.module_webutils.service_ekey.service;
 
 public interface UtilsWebChatService {
     void registerUtilsKeyModuleService();
